feat: сайдменю и исправления верстки
