+++ src/components/Main/components/Portfolio/Portfolio.jsx
@@ -4,13 +4,13 @@ const Portfolio = () => (
   <section className="portfolio">
     <header className="portfolio__header">Портфолио</header>
     <section className="portfolio__links">
-      <a className="portfolio__link" href="https://github.com/am-avraam/blog-platform">
+      <a className="portfolio__link" href="https://github.com/am-avraam/mesto">
         Статичный сайт
       </a>
       <a className="portfolio__link" href="https://github.com/am-avraam/blog-platform">
         Адаптивный сайт
       </a>
-      <a className="portfolio__link" href="https://github.com/am-avraam/blog-platform">
+      <a className="portfolio__link" href="https://github.com/am-avraam/react-mesto-api-full-gha">
         Одностраничное приложение
       </a>
     </section>
